feat: add implement flag banner on feature flag overview
Show a banner prompting users to wrap their feature logic in a flag
evaluation when the project SDK is connected but no metrics have been
received for this specific flag yet. Extracts a shared banner card
used by both the connect SDK and implement flag banners. Gated behind
the onboardingFlagSetup feature flag.
